requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy[asyncio]==2.0.32
asyncpg==0.29.0
python-dotenv==1.0.1
elasticsearch==8.14.0

aiokafka==0.10.0
debugpy==1.8.*
kafka-python==2.0.2
